+++ underthesea/underthesea.py
@@ -1,11 +1,6 @@
 # -*- coding: utf-8 -*-
-def hi():
-    print("Hello")
-
-
 class Corpus:
-    """
-    Corpus is fundamental resource of NLP
+    """Corpus is fundamental resource of NLP
     """
     def __init__(self):
         print("Corpus is fundamental resource of NLP")